# backend/cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Healthcare platform API with FHIR R4 interoperability",
        "title": "Healthcare API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@healthcare.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/encounters/{encounterFhirId}/observations": {
            "get": {
                "description": "Returns all observations/vital signs for an encounter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "observations"
                ],
                "summary": "List observations by encounter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encounter FHIR ID",
                        "name": "encounterFhirId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/observation.ObservationResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new observation/vital sign for an encounter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "observations"
                ],
                "summary": "Create an observation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encounter FHIR ID",
                        "name": "encounterFhirId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Observation data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/observation.CreateObservationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/observation.CreateObservationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/patients/{patientFhirId}/observations": {
            "get": {
                "description": "Returns all observations/vital signs for a patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "observations"
                ],
                "summary": "List observations by patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient FHIR ID",
                        "name": "patientFhirId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/observation.ObservationResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "observation.CreateObservationRequest": {
            "type": "object",
            "properties": {
                "code_display": {
                    "type": "string"
                },
                "loinc_code": {
                    "type": "string"
                },
                "patient_fhir_id": {
                    "type": "string"
                },
                "value_quantity": {
                    "type": "number"
                },
                "value_unit": {
                    "type": "string"
                }
            }
        },
        "observation.CreateObservationResponse": {
            "type": "object",
            "properties": {
                "code_display": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "encounter_fhir_id": {
                    "type": "string"
                },
                "fhir_id": {
                    "type": "string"
                },
                "loinc_code": {
                    "type": "string"
                },
                "patient_fhir_id": {
                    "type": "string"
                },
                "value_quantity": {
                    "type": "number"
                },
                "value_unit": {
                    "type": "string"
                }
            }
        },
        "observation.ObservationResponse": {
            "type": "object",
            "properties": {
                "code_display": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "encounter_fhir_id": {
                    "type": "string"
                },
                "fhir_id": {
                    "type": "string"
                },
                "loinc_code": {
                    "type": "string"
                },
                "patient_fhir_id": {
                    "type": "string"
                },
                "value_quantity": {
                    "type": "number"
                },
                "value_unit": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}